veloapp/requirements.txt
django
djangorestframework
psycopg